updating

+++ src/Day_005/MultiSig.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.19;
 
 contract multiSigWallet {
-    error NotEnoughApprovals(uint required, uint current);
+    error NotEnoughApprovals(uint256 required, uint256 current);
     error AlreadyExecuted();
     error AlreadyApproved();
     error InvalidRecipient();
@@ -10,15 +10,10 @@ contract multiSigWallet {
     error OwnerDuplicated();
     error OwnerCantBeZero();
 
-    event TransactionSubmitted(
-        address indexed sender,
-        uint indexed txId,
-        address to,
-        uint amount
-    );
+    event TransactionSubmitted(address indexed sender, uint256 indexed txId, address to, uint256 amount);
 
-    event TransactionExecuted(uint indexed txId, address to, uint amount);
-    event TransactionApproved(uint indexed txId, address indexed owner);
+    event TransactionExecuted(uint256 indexed txId, address to, uint256 amount);
+    event TransactionApproved(uint256 indexed txId, address indexed owner);
 
     uint256 requiredApprovals = 2;
     uint256 txIdCounter;
@@ -42,7 +37,7 @@ contract multiSigWallet {
     }
 
     constructor(address[3] memory _owners) {
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             address owner = _owners[i];
             if (owner == address(0)) {
                 revert OwnerCantBeZero();
@@ -58,19 +53,14 @@ contract multiSigWallet {
 
     function submitTransaction(address to, uint256 amount) external onlyOwner {
         uint256 txId = txIdCounter;
-        transactions[txId] = Transaction({
-            to: to,
-            amount: amount,
-            executed: false,
-            approvals: 0
-        });
+        transactions[txId] = Transaction({to: to, amount: amount, executed: false, approvals: 0});
 
         txIdCounter += 1;
 
         emit TransactionSubmitted(msg.sender, txId, to, amount);
     }
 
-    function approveTransaction(uint txId) external onlyOwner {
+    function approveTransaction(uint256 txId) external onlyOwner {
         Transaction storage txn = transactions[txId];
         if (hasApproved[txId][msg.sender]) {
             revert AlreadyApproved();
@@ -89,7 +79,7 @@ contract multiSigWallet {
         emit TransactionApproved(txId, msg.sender);
     }
 
-    function executeTransaction(uint txId) external onlyOwner {
+    function executeTransaction(uint256 txId) external onlyOwner {
         Transaction storage txn = transactions[txId];
         if (txn.executed) {
             revert AlreadyExecuted();
@@ -105,7 +95,7 @@ contract multiSigWallet {
 
         txn.executed = true;
 
-        (bool success, ) = payable(txn.to).call{value: txn.amount}("");
+        (bool success,) = payable(txn.to).call{value: txn.amount}("");
         require(success, "Transfer failed");
 
         emit TransactionExecuted(txId, txn.to, txn.amount);
